rbac_project/requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
Django==5.1.3
django-cors-headers==4.6.0
django-dotenv==1.4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
PyJWT==2.10.1
sqlparse==0.5.2
